\name{igdbV4-package}
\alias{igdbV4-package}
\alias{igdbV4}
\docType{package}
\title{
\packageTitle{igdbV4}
}
\description{
\packageDescription{igdbV4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{igdbV4}
\packageIndices{igdbV4}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{igdbV4}

Maintainer: \packageMaintainer{igdbV4}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
